Changed logic in EmployeeCourseList
