+++ frontend/src/pages/EmployeeProfile/EmployeeSkillsAndCourses/EmployeeCourses/EmployeeCoursesContainer.tsx
@@ -10,7 +10,6 @@ import EmployeeCourses from './EmployeeCourses';
 const EmployeeCoursesContainer: React.FC<IEmployeeCoursesList> = ({ courses }) => {
   const [searchCourse, setSearchCourse] = useState('');
   const [employeeCourses, setEmployeeCourses] = useState<IClientCourse[]>();
-  const [isShown, setIsShown] = useState(false);
 
   useEffect(() => {
     if (courses) {
@@ -36,20 +35,9 @@ const EmployeeCoursesContainer: React.FC<IEmployeeCoursesList> = ({ courses }) =
     checkWhitespace(event, searchCourse);
   };
 
-  const showCourseInfo = () => {
-    setIsShown(true);
-  };
-
-  const hideCourseInfo = () => {
-    setIsShown(false);
-  };
-
   return (
     <EmployeeCourses
       courses={employeeCourses}
-      isShown={isShown}
-      showCourseInfo={showCourseInfo}
-      hideCourseInfo={hideCourseInfo}
       searchCourseInList={searchCourseInList}
       checkSpace={checkSpace}
       checkPastedValue={checkPastedValue}
