+++ frontend/src/types/employee.ts
@@ -26,16 +26,10 @@ export interface IEmployeeCourses {
   checkPastedValue: (event: React.ClipboardEvent) => void;
   searchCourse: string;
   courses?: IClientCourse[];
-  isShown?: boolean;
-  showCourseInfo?: () => void;
-  hideCourseInfo?: () => void;
 }
 
 export interface IEmployeeCoursesList {
   courses?: IClientCourse[];
-  isShown?: boolean;
-  showCourseInfo?: () => void;
-  hideCourseInfo?: () => void;
 }
 
 export interface IEmployeeInfo {
